Change Interop.SizeOf to return long instead of int

+++ CSGL.Vulkan/Vulkan/CommandBuffer.cs
@@ -350,7 +350,7 @@ namespace CSGL.Vulkan {
         public void PushConstants<T>(PipelineLayout layout, VkShaderStageFlags stageFlags, uint offset, List<T> data) where T : struct {
             unsafe
             {
-                int size = Interop.SizeOf<T>();
+                int size = (int)Interop.SizeOf<T>();
                 byte* native = stackalloc byte[size * data.Count];
                 Interop.Copy(data, (IntPtr)native);
                 Device.Commands.cmdPushConstants(commandBuffer, layout.Native, stageFlags, offset, (uint)(size * data.Count), (IntPtr)native);
